separate models into their own files
